Postman tests in process.

+++ ewm-main-service/src/main/java/ru/practicum/ewm/APIadmin/service/AdminCategoryServiceImpl.java
@@ -1,21 +1,27 @@
 package ru.practicum.ewm.APIadmin.service;
 
+import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.ewm.common.dto.CategoryDto;
 import ru.practicum.ewm.common.dto.NewCategoryDto;
 import ru.practicum.ewm.common.exception.BadRequestException;
+import ru.practicum.ewm.common.exception.ForbiddenException;
 import ru.practicum.ewm.common.exception.NotFoundException;
 import ru.practicum.ewm.common.mapper.CategoryMapper;
 import ru.practicum.ewm.common.model.Category;
 import ru.practicum.ewm.common.repository.CategoryRepository;
+import ru.practicum.ewm.common.repository.EventRepository;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class AdminCategoryServiceImpl implements AdminCategoryService {
-    private final CategoryRepository repository;
+    CategoryRepository repository;
+    EventRepository eventRepository;
 
     @Transactional
     @Override
@@ -27,7 +33,11 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
     @Override
     public void delete(Long catId) {
         getByIdWithCheck(catId);
-        repository.deleteById(catId);
+        if (eventRepository.findAllByCategoryId(catId).size() == 0) {
+            repository.deleteById(catId);
+        } else {
+            throw new ForbiddenException("The category is not empty");
+        }
     }
 
     @Transactional
